refactor(benchmark)!: remove the external agent execution bridge
